GEODE-4154: Removed GemFireCacheImpl.getInstance
	* Refactored the test code so that GemFireCacheImpl.getInstance is nolonger called.
	* Upgraded the WANTestBase to use LogService than LoggerServiceUtils.
	* Optimized the imports.
